[refactor] 코드 도메인 폴더 단위 분리
