+++ src/main/java/com/oreo/finalproject_5re5_be/code/exception/CodeNotFoundException.java
@@ -1,4 +1,4 @@
-package com.oreo.finalproject_5re5_be.member.exception;
+package com.oreo.finalproject_5re5_be.code.exception;
 
 import static com.oreo.finalproject_5re5_be.global.exception.ErrorCode.*;
