+++ src/main/java/com/oreo/finalproject_5re5_be/member/entity/MemberDelete.java
@@ -1,8 +1,8 @@
 package com.oreo.finalproject_5re5_be.member.entity;
 
+import com.oreo.finalproject_5re5_be.code.entity.Code;
 import com.oreo.finalproject_5re5_be.global.entity.BaseEntity;
 import com.oreo.finalproject_5re5_be.member.dto.request.MemberRemoveRequest;
-import com.oreo.finalproject_5re5_be.member.repository.MemberDeleteRepository;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -11,7 +11,6 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
