+++ src/test/java/com/oreo/finalproject_5re5_be/member/repository/MemberChangeHistoryRepositoryTest.java
@@ -2,7 +2,8 @@ package com.oreo.finalproject_5re5_be.member.repository;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import com.oreo.finalproject_5re5_be.member.entity.Code;
+import com.oreo.finalproject_5re5_be.code.entity.Code;
+import com.oreo.finalproject_5re5_be.code.repository.CodeRepository;
 import com.oreo.finalproject_5re5_be.member.entity.Member;
 import com.oreo.finalproject_5re5_be.member.entity.MemberChangeHistory;
 import java.time.LocalDateTime;
