+++ src/test/java/com/oreo/finalproject_5re5_be/member/service/CodeServiceImplTest.java
@@ -4,13 +4,14 @@ import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.when;
 
-import com.oreo.finalproject_5re5_be.member.dto.request.CodeRequest;
-import com.oreo.finalproject_5re5_be.member.dto.request.CodeUpdateRequest;
-import com.oreo.finalproject_5re5_be.member.dto.response.CodeResponse;
-import com.oreo.finalproject_5re5_be.member.dto.response.CodeResponses;
-import com.oreo.finalproject_5re5_be.member.entity.Code;
-import com.oreo.finalproject_5re5_be.member.exception.CodeDuplicatedException;
-import com.oreo.finalproject_5re5_be.member.repository.CodeRepository;
+import com.oreo.finalproject_5re5_be.code.dto.request.CodeRequest;
+import com.oreo.finalproject_5re5_be.code.dto.request.CodeUpdateRequest;
+import com.oreo.finalproject_5re5_be.code.dto.response.CodeResponse;
+import com.oreo.finalproject_5re5_be.code.dto.response.CodeResponses;
+import com.oreo.finalproject_5re5_be.code.entity.Code;
+import com.oreo.finalproject_5re5_be.code.exception.CodeDuplicatedException;
+import com.oreo.finalproject_5re5_be.code.repository.CodeRepository;
+import com.oreo.finalproject_5re5_be.code.service.CodeServiceImpl;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
